Add files via upload
This is the Icestudio compatible Collection for signed integer arithmetic. Focus is on 16 and 24 bit functions for +, - , *,  / for DSP or control type of applications - hence the saturation option for most blocks. Also has some 32 and 48 bit blocks, and scaling with a multiply and then divide in the same block, using internally twice the word sizes. The multiply, divide and square root blocks need a clock and a trigger because they're serial implementations. The plus minus min max are implemented without a clock so do not need a trigger.